+++ litert/vendors/qualcomm/context_binary_info.h
@@ -21,11 +21,10 @@
 
 #include "litert/cc/litert_expected.h"
 #include "litert/vendors/qualcomm/core/wrappers/tensor_wrapper.h"
+#include "litert/vendors/qualcomm/qnn_manager.h"
 #include "System/QnnSystemContext.h"  // from @qairt
 
 namespace litert::qnn {
-
-class QnnManager;
 
 class GraphInfo {
  public:
