+++ litert/vendors/qualcomm/qnn_api_loader.cc
@@ -5,13 +5,10 @@
 
 #include <stdlib.h>
 
-#include <charconv>
 #include <cstdint>
 #include <memory>
 #include <optional>
 #include <string>
-#include <string_view>
-#include <system_error>
 #include <utility>
 
 #include "absl/strings/str_cat.h"  // from @com_google_absl
@@ -43,46 +40,6 @@ static constexpr int kRequiredNumProviders{1};
 }
 
 namespace litert::qnn {
-
-Expected<SdkVersion> ParseSdkVersion(const char* build_id) {
-  // Generic parse-failure result.
-  const auto parsing_error =
-      Unexpected(kLiteRtStatusErrorRuntimeFailure, "Failed to parse build ID");
-
-  if (!build_id) return parsing_error;
-
-  std::string_view version_str = build_id;
-
-  // Require the 'v' prefix, then strip it.
-  if (version_str.empty() || version_str.front() != 'v') {
-    return parsing_error;
-  }
-  version_str.remove_prefix(1);
-
-  SdkVersion version{};
-  const char* current = version_str.data();
-  const char* const end = version_str.data() + version_str.size();
-
-  auto parse_component = [&current, &end](int& component) {
-    auto [ptr, ec] = std::from_chars(current, end, component);
-    if (ec != std::errc()) {
-      return false;
-    }
-    current = ptr;
-    return true;
-  };
-
-  // Expect "major.minor.patch".
-  if (!parse_component(version.major)) return parsing_error;
-
-  if (current == end || *current++ != '.') return parsing_error;
-  if (!parse_component(version.minor)) return parsing_error;
-
-  if (current == end || *current++ != '.') return parsing_error;
-  if (!parse_component(version.patch)) return parsing_error;
-
-  return version;
-}
 
 namespace {
 
@@ -490,7 +447,7 @@ LiteRtStatus QnnApiLoader::LoadLibraries(
     LITERT_LOG(LITERT_ERROR, "%s", "Failed to get QNN backend build ID");
     return kLiteRtStatusErrorRuntimeFailure;
   }
-  LITERT_ASSIGN_OR_RETURN(sdk_version_, ParseSdkVersion(build_id));
+  build_id_ = build_id;
 
   return kLiteRtStatusOk;
 }
