+++ litert/vendors/qualcomm/qnn_manager.cc
@@ -18,10 +18,13 @@
 #include "litert/vendors/qualcomm/qnn_manager.h"
 
 #include <array>
+#include <charconv>
 #include <cstdint>
 #include <memory>
 #include <optional>
 #include <string>
+#include <string_view>
+#include <system_error>
 #include <utility>
 #include <vector>
 
@@ -139,12 +142,53 @@ Expected<QnnManager> QnnManager::Create(const QnnApiLoader& loader,
       return Unexpected(kLiteRtStatusErrorRuntimeFailure,
                         "Unsupported backend type");
   }
+  LITERT_ASSIGN_OR_RETURN(SdkVersion sdk_version,
+                          ParseSdkVersion(loader.GetBuildId().c_str()));
   QnnManager qnn_manager(loader.Api(), loader.SystemApi(), &options,
-                         loader.GetSdkVersion(), std::move(backend), bound_soc,
-                         mode);
+                         sdk_version, std::move(backend), bound_soc, mode);
   // Fresh backend handle -- apply the configured custom op package.
   LITERT_RETURN_IF_ERROR(RegisterCustomOpPackage(qnn_manager, mode));
   return qnn_manager;
+}
+
+Expected<SdkVersion> QnnManager::ParseSdkVersion(const char* build_id) {
+  // Generic parse-failure result.
+  const auto parsing_error =
+      Unexpected(kLiteRtStatusErrorRuntimeFailure, "Failed to parse build ID");
+
+  if (!build_id) return parsing_error;
+
+  std::string_view version_str = build_id;
+
+  // Require the 'v' prefix, then strip it.
+  if (version_str.empty() || version_str.front() != 'v') {
+    return parsing_error;
+  }
+  version_str.remove_prefix(1);
+
+  SdkVersion version{};
+  const char* current = version_str.data();
+  const char* const end = version_str.data() + version_str.size();
+
+  auto parse_component = [&current, &end](int& component) {
+    auto [ptr, ec] = std::from_chars(current, end, component);
+    if (ec != std::errc()) {
+      return false;
+    }
+    current = ptr;
+    return true;
+  };
+
+  // Expect "major.minor.patch".
+  if (!parse_component(version.major)) return parsing_error;
+
+  if (current == end || *current++ != '.') return parsing_error;
+  if (!parse_component(version.minor)) return parsing_error;
+
+  if (current == end || *current++ != '.') return parsing_error;
+  if (!parse_component(version.patch)) return parsing_error;
+
+  return version;
 }
 
 LiteRtStatus QnnManager::ValidateOp(::qnn::OpWrapper& op) {
